表达式计算

+++ software/expression.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 // º¯ÊıÉùÃ÷
-double expression_calc(uint8_t item[], uint8_t n);
+float expression_calc(uint8_t item[], uint8_t n);
 
 #endif // EXPRESSION_H
